Fix bug with shoving stuff at the end of a file
The fix is to have a separate file for the source and sink, and then
rename one to the other at the end. I had been tempted to not do this to
start with, because I didn't want to get into a spot where the
permissions / owner / group / created date could be wrong, but it turns
out:

- proper unix doesn't have a concept of "created date" so i felt more
  justified to overwrite the file
- I looked at how freebsd sed does it, and it just does the chown /
  chmod
- I was already inheriting the permissions from the original file

So given those three things I decided to bite the bullet and just use a
tempfile.